backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
openai==1.35.0
anthropic==0.28.0
boto3==1.34.0
python-dotenv==1.0.1
python-multipart==0.0.9
